@@ -23,9 +23,10 @@ dependencies:
   yaml: ^3.1.0
 
 dev_dependencies:
+  build_runner: ^2.0.3
   matcher: ^0.12.10
   mockito: ^5.0.7
   pedantic: ^1.11.0
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
